Implement UDPlz4 Remote variant
This variant will compress and de-compress bytes before and after sending
on the UDP sockets.

In best case this can reduce used bandwidth by ~90% if the tunneled data
is very compressible. If the data is not compressible, however the overhead
of lz4 is negligible.

This commit also includes a netem test environment, pre-configured for testing
this new functionality and which will serve as an example of usage.