Delegate the responsibility for keep alive to Remote
How keepalive is performed can vary depending on the
transport protocol used for a Remote. The UDP remote
needs manual handling of keepalive while a QUIC based
remote will have it built in as part of the protocol.

Therefore, keepalive should be handled by each
individual Remote instead of begin inside the main
event loop.

This commit delegates the keepalive to the Remote
which will handle the transmission of keepalive
messages however it sees fit.