Allow specifying peers as a list in settings file
This commit changes the settings file to allow specifying
the peers as a list of Socketaddrs.

The list must be specified and can contain 0 or more peers.
If 0 peers are specified, mpconn will not be able to transmit
any packets on the remote, unless peers has been auto discovered
through keepalive messages.

The example configurations in test_tools have been updated to match
this change.

WARNING this commit contains breaking changes to the settings format