Add reordering Sequencer module:
The Sequencer uses a B-tree of packets wrapped in a struct with a timestamp.
Received packets are inserted in the B-tree and the sequence number is tracked
by the Sequencer.

If we have a whole in the sequence number we can prune packets after a
configurable deadline. This ensures that we never wait longer than necessary