Add internal_messages
In order to make it easier to share information between
different sections of the code, we should make internal messages
which wrap the actual messages with additional helpful info.

In this commit we begin by wrapping all incomming Packets in an
IncommingPacket struct which contains the receiver interface name,
and the address from which the packet was received.

This makes it possible to better log where we received a given packet.