Introduce a remote_manager:
This is an attempt at moving the whole architecture over to message passing
in order to decouple the various parts of the design.

Instead of having the Remotes in the top level main event loop, we now run
them inside a remote_manager and pass/receive pakcets through Broadcast- and
mpsc-channels.